refactor: Answer 을 view 패키지로 이동시킨다.
