+++ src/main/java/view/Answer.java
@@ -1,4 +1,4 @@
-package domain;
+package view;
 
 import java.util.Arrays;
